[MIG] Forward port of lp:~therp-nl/banking-addons/6.1-link_partner_wizard

+++ account_banking/wizard/banking_transaction_wizard.py
@@ -73,7 +73,7 @@ class banking_transaction_wizard(orm.TransientModel):
         Call the automatic matching routine for one or
         more bank transactions
         """
-        if isinstance(ids, (int, float)):
+        if isinstance(ids, (int, long)):
             ids = [ids]
         import_transaction_obj = self.pool.get('banking.import.transaction')
         trans_id = self.read(
@@ -276,7 +276,7 @@ class banking_transaction_wizard(orm.TransientModel):
         settings_pool = self.pool.get('account.banking.account.settings')
         statement_pool = self.pool.get('account.bank.statement.line')
 
-        if isinstance(ids, (int, float)):
+        if isinstance(ids, (int, long)):
             ids = [ids]
 
         for wiz in self.browse(cr, uid, ids, context=context):
@@ -309,7 +309,7 @@ class banking_transaction_wizard(orm.TransientModel):
         return self.create_act_window(cr, uid, ids, context=None)
 
     def reverse_duplicate(self, cr, uid, ids, context=None):
-        if isinstance(ids, (int, float)):
+        if isinstance(ids, (int, long)):
             ids = [ids]
         transaction_obj = self.pool.get('banking.import.transaction')
         for wiz in self.read(
